Initial Commit
This is the current state of the code as I make my first commit to GitHub. The test_nixnet_can_dll.py and nixnet_custom.py are the two main files I have modified. At this stage, I am successfully starting an xnet session and reading a raw frame, but not much else.